src/producer/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o src/producer/requirements.txt
certifi==2024.7.4
    # via requests
charset-normalizer==3.3.2
    # via requests
idna==3.7
    # via requests
kafka-python-ng==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
    # via python-dateutil
tzdata==2024.1
urllib3==2.2.2
    # via requests
